move WifiCommon to wifi module
